refactor: DTOディレクトリの再構成
## 変更内容
- controllers/dto/ から独立した dto/ ディレクトリに移動
- リクエスト/レスポンスの分離構造に変更
- 自動生成ファイルの配置を整理

## 新しい構造
backend/dto/
├── requests/          # リクエストDTO
│   └── project.go     # プロジェクト関連リクエスト
├── responses/         # レスポンスDTO
│   └── project.go     # プロジェクト関連レスポンス
└── generated/         # 自動生成されたDTO
    ├── types.go       # 型定義
    ├── server.go      # サーバーコード
    └── spec.go        # バリデーション

## 改善点
- 責任の明確化: DTO層とコントローラー層の分離
- 再利用性の向上: 複数のコントローラーで使用可能
- 拡張性の向上: 新しいAPIエンドポイントの追加が容易
- テスト容易性: DTOを独立してテスト可能

## 更新されたファイル
- Makefile: 生成パスの更新
- scripts/generate-dto.sh: 出力ディレクトリの変更
- backend/dto/: 新しいDTO構造

これにより、より適切で保守しやすいアーキテクチャになりました。
