+++ scripts/generate-dto.sh
@@ -14,38 +14,38 @@ if ! command -v oapi-codegen &> /dev/null; then
 fi
 
 # 出力ディレクトリの作成
-mkdir -p backend/controllers/dto
+mkdir -p backend/dto/generated
 
 # DTO型の生成
 echo "📝 DTO型を生成しています..."
 oapi-codegen \
-    -package dto \
+    -package generated \
     -generate types \
-    openapi.yaml > backend/controllers/dto/generated_types.go
+    openapi.yaml > backend/dto/generated/types.go
 
 # バリデーション関数の生成
 echo "🔍 バリデーション関数を生成しています..."
 oapi-codegen \
-    -package dto \
+    -package generated \
     -generate spec \
-    openapi.yaml > backend/controllers/dto/generated_spec.go
+    openapi.yaml > backend/dto/generated/spec.go
 
 # サーバーコードの生成（オプション）
 echo "🖥️ サーバーコードを生成しています..."
 oapi-codegen \
-    -package dto \
+    -package generated \
     -generate server \
-    openapi.yaml > backend/controllers/dto/generated_server.go
+    openapi.yaml > backend/dto/generated/server.go
 
 echo "✅ DTO生成が完了しました！"
 echo "📁 生成されたファイル:"
-echo "   - backend/controllers/dto/generated_types.go"
-echo "   - backend/controllers/dto/generated_spec.go"
-echo "   - backend/controllers/dto/generated_server.go"
+echo "   - backend/dto/generated/types.go"
+echo "   - backend/dto/generated/spec.go"
+echo "   - backend/dto/generated/server.go"
 
 # 生成されたコードの確認
 echo "🔍 生成されたコードを確認しています..."
-go fmt backend/controllers/dto/*.go
-go vet backend/controllers/dto/*.go
+go fmt backend/dto/generated/*.go
+go vet backend/dto/generated/*.go
 
 echo "🎉 すべて完了しました！"
